+++ src/CMakeLists.txt
@@ -2,7 +2,6 @@
 # Include our other top level modules
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Engine)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Glad)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Shader)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Graphics)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Utility)
 
@@ -11,7 +10,6 @@ set(SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/dde.cpp
     ${ENGINE_SOURCES}
     ${GLAD_SOURCES}
-    ${SHADER_SOURCES}
     ${GRAPHICS_SOURCES}
     ${UTILITY_SOURCES}
     PARENT_SCOPE)
