+++ src/Graphics/CMakeLists.txt
@@ -1,6 +1,7 @@
 # Append the src files to SOURCES
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Vertex)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Shape)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Shader)
 
 set(GRAPHICS_SOURCES 
   ${CMAKE_CURRENT_SOURCE_DIR}/Sprite.cpp 
@@ -9,4 +10,5 @@ set(GRAPHICS_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/Pencil.cpp
   ${VERTEX_SOURCES}
   ${SHAPE_SOURCES}
+  ${SHADER_SOURCES}
   PARENT_SCOPE)
