+++ src/Graphics/Shader/ShaderStore.cpp
@@ -1,6 +1,6 @@
 #include <Glad/glad/glad.h>
 
-#include <DDE/Shader/ShaderStore.hpp>
+#include <DDE/Graphics/Shader/ShaderStore.hpp>
 #include <DDE/Utility/ShaderTypes.hpp>
 #include <iostream>
 #include <stdexcept>
@@ -43,23 +43,23 @@ unsigned int DDE::ShaderStore::loadShaderPipeline(DDE::ShaderStage stage) {
  */
 void DDE::ShaderStore::_cacheShaderPipeline(DDE::ShaderStage stage) {
   switch (stage) {
-    case DDE::ShaderStage::PRIMITIVE:
-      this->_pipelineMappings.insert(
-          {stage,
-           DDE::ShaderProgram({DDE::Shader("../glsl/primitive/primitive.vert",
-                                           GL_VERTEX_SHADER),
-                               DDE::Shader("../glsl/primitive/primitive.frag",
-                                           GL_FRAGMENT_SHADER)})});
-      break;
-    case DDE::ShaderStage::SCENE:
-      throw std::runtime_error("Not Implemented!\n");
-      break;
-    case DDE::ShaderStage::OBJECT:
-      throw std::runtime_error("Not Implemented!\n");
-      break;
-    default:
-      throw std::runtime_error(
-          "Invalid Shader Stage was Provided to the Shader Engine!\n");
+  case DDE::ShaderStage::PRIMITIVE:
+    this->_pipelineMappings.insert(
+        {stage,
+         DDE::ShaderProgram(
+             {DDE::Shader("../glsl/primitive/primitive.vert", GL_VERTEX_SHADER),
+              DDE::Shader("../glsl/primitive/primitive.frag",
+                          GL_FRAGMENT_SHADER)})});
+    break;
+  case DDE::ShaderStage::SCENE:
+    throw std::runtime_error("Not Implemented!\n");
+    break;
+  case DDE::ShaderStage::OBJECT:
+    throw std::runtime_error("Not Implemented!\n");
+    break;
+  default:
+    throw std::runtime_error(
+        "Invalid Shader Stage was Provided to the Shader Engine!\n");
   }
 }
