+++ src/dde.cpp
@@ -1,14 +1,14 @@
-#include "DDE/Graphics/Shape/Quad.hpp"
+#include <DDE/Graphics/Shape/Quad.hpp>
 #include <Glad/glad/glad.h>
 
 #include <DDE/Engine/RenderEngine.hpp>
 #include <DDE/Engine/ShaderEngine.hpp>
 #include <DDE/Graphics/Pencil.hpp>
+#include <DDE/Graphics/Shader/Shader.hpp>
+#include <DDE/Graphics/Shader/ShaderProgram.hpp>
 #include <DDE/Graphics/Shape/Triangle.hpp>
 #include <DDE/Graphics/Sprite.hpp>
 #include <DDE/Graphics/Vertex/Vertex.hpp>
-#include <DDE/Shader/Shader.hpp>
-#include <DDE/Shader/ShaderProgram.hpp>
 
 void drawFunction(DDE::Pencil &pencil, DDE::Quad &square,
                   DDE::Triangle &triange) {
